requirements/base.txt
dagster
dagster-webserver
dagster-dbt
dbt-core
dbt-bigquery
fastapi
uvicorn[standard]
pydantic-settings
google-cloud-bigquery
google-cloud-storage
google-cloud-secret-manager
google-cloud-logging
google-cloud-monitoring
psycopg[binary]
sqlalchemy
httpx
polars
pyarrow
python-dotenv
sodapy
soda-core
